config/config_variables.py
# config/config_variables.py

import os
from dotenv import load_dotenv

# Cargar variables del .env
load_dotenv()

class Settings:
    DB_USER: str = os.getenv("DB_USER", "default_user")
    DB_PASSWORD: str = os.getenv("DB_PASSWORD", "default_password")
    DB_HOST: str = os.getenv("DB_HOST", "localhost")
    DB_NAME: str = os.getenv("DB_NAME", "test_db")

settings = Settings()