requirements.txt
python-dotenv==1.0.1
sqlalchemy==2.0.25
pymysql==1.1.0
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
